backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
langchain
langchain-groq
langchain-community
duckduckgo-search
langgraph
PyJWT
passlib[bcrypt]
python-multipart
